+++ libawui/awui/Emulation/Chip8/Memory.h
@@ -21,8 +21,8 @@ namespace awui {
 
 					void LoadRom(const String file);
 
-					uint8_t ReadByte(int pos);
-					void WriteByte(int pos, uint8_t value);
+					uint8_t ReadByte(int64_t pos);
+					void WriteByte(int64_t pos, uint8_t value);
 
 					void Reload();
 			};
