+++ libawui/awui/Emulation/Chip8/Registers.h
@@ -10,7 +10,7 @@ namespace awui {
 				private:
 					uint8_t * _v;
 					uint8_t _length;
-					uint16_t _i;
+					uint32_t _i;
 
 				public:
 					Registers(uint8_t n);
@@ -21,8 +21,8 @@ namespace awui {
 					void SetV(uint8_t pos, uint8_t value);
 					uint8_t GetV(uint8_t pos);
 
-					void SetI(uint16_t value);
-					uint16_t GetI();
+					void SetI(uint32_t value);
+					uint32_t GetI();
 			};
 		}
 	}
